toolkit/components/url-classifier/tests/mochitest/test_classifier.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test the URI Classifier</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="classifierHelper.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>

<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">

<script class="testbody" type="text/javascript">

var firstLoad = true;

// Add some URLs to the malware database.
var testData = [
  { url: "malware.example.com/",
    db: "test-malware-simple"
  },
  { url: "unwanted.example.com/",
    db: "test-unwanted-simple"
  }
];

function loadTestFrame() {
  document.getElementById("testFrame").src = "classifierFrame.html";
}

// Expected finish() call is in "classifierFrame.html".
SimpleTest.waitForExplicitFinish();

function updateSuccess() {
  SpecialPowers.pushPrefEnv(
    {"set" : [["browser.safebrowsing.malware.enabled", true]]},
    loadTestFrame);
}

function updateError(errorCode) {
  ok(false, "Couldn't update classifier. Error code: " + errorCode);
  // Abort test.
  SimpleTest.finish();
}

SpecialPowers.pushPrefEnv(
  {"set" : [["urlclassifier.malwareTable", "test-malware-simple,test-unwanted-simple"],
            ["urlclassifier.phishTable", "test-phish-simple"]]},
  function() {
    classifierHelper.addUrlToDB(testData)
      .then(updateSuccess)
      .catch(err => {
        updateError(err);
      });
  });

</script>

</pre>
<iframe id="testFrame" onload=""></iframe>
</body>
</html>
